+++ src/OpenTelemetry.OpAmp.Client/Internal/Listeners/IOpAmpMessage.cs
@@ -1,0 +1,8 @@
+// Copyright The OpenTelemetry Authors
+// SPDX-License-Identifier: Apache-2.0
+
+namespace OpenTelemetry.OpAmp.Client.Internal.Listeners;
+
+internal interface IOpAmpMessage
+{
+}