+++ src/OpenTelemetry.OpAmp.Client/Listeners/IOpAmpMessage.cs
@@ -1,8 +1,0 @@
-// Copyright The OpenTelemetry Authors
-// SPDX-License-Identifier: Apache-2.0
-
-namespace OpenTelemetry.OpAmp.Client.Listeners;
-
-internal interface IOpAmpMessage
-{
-}