+++ test/OpenTelemetry.OpAmp.Client.Tests/FrameDispatcherTests.cs
@@ -3,6 +3,7 @@
 
 #if NET
 
+using OpenTelemetry.OpAmp.Client.Internal;
 using OpenTelemetry.OpAmp.Client.Tests.Mocks;
 using Xunit;