+++ test/OpenTelemetry.Resources.AWS.Tests/AWSEC2DetectorTests.cs
@@ -1,20 +1,50 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
+#if NETFRAMEWORK
+using System.Net.Http;
+#endif
 using Xunit;
 
 namespace OpenTelemetry.Resources.AWS.Tests;
 
 public class AWSEC2DetectorTests
 {
+    private const string AwsEc2InstanceIdUrl = "http://169.254.169.254/latest/meta-data/instance-id";
+    private static readonly HttpClient HttpClient = new()
+    {
+        Timeout = TimeSpan.FromSeconds(3),
+    };
+
+    public static async Task<bool> IsRunningOnEC2()
+    {
+        try
+        {
+            var instanceId = await HttpClient.GetStringAsync(new Uri(AwsEc2InstanceIdUrl));
+            return !string.IsNullOrEmpty(instanceId);
+        }
+        catch
+        {
+            return false;
+        }
+    }
+
     [Fact]
-    public void TestDetect()
+    public async Task TestDetect()
     {
         var awsEC2Detector = new AWSEC2Detector(
             new OpenTelemetry.AWS.AWSSemanticConventions(
                 SemanticConventionVersion.Latest));
 
-        Assert.Empty(awsEC2Detector.Detect().Attributes); // will be null as it's not in ec2 environment
+        var attributes = awsEC2Detector.Detect().Attributes;
+        if (!await IsRunningOnEC2())
+        {
+            Assert.Empty(attributes); // will be null as it's not in ec2 environment
+        }
+        else
+        {
+            Assert.NotEmpty(attributes);
+        }
     }
 
     [Fact]